Added HealthBar class

+++ Rogue_DianaAmaro103678/src/pt/iscte/poo/example/Moveable.java
@@ -7,13 +7,13 @@ import pt.iscte.poo.utils.Vector2D;
 
 public class Moveable extends AbstractObject {
 
-    private double hitPoints = 0;
+    private int hitPoints = 0;
 
     public Moveable(String name, Point2D position) {
         super(name, position);
     }
 
-    public Moveable(String name, Point2D position, double hitPoints) {
+    public Moveable(String name, Point2D position, int hitPoints) {
         super(name, position);
         this.hitPoints = hitPoints;
     }
@@ -30,15 +30,15 @@ public class Moveable extends AbstractObject {
         return 1;
     }
 
-    public double getHitPoints() {
+    public int getHitPoints() {
         return hitPoints;
     }
 
-    public void setHitPoints(double hitPoints) {
+    public void setHitPoints(int hitPoints) {
         this.hitPoints = hitPoints;
     }
 
-    public void updateHitPoints(double delta) {
+    public void updateHitPoints(int delta) {
         this.hitPoints = this.hitPoints + delta;
     }
 }
